type:feature scope: comment:文件传输.

+++ lib-clink/src/main/java/box/AbsByteArrayReceivePacket.java
@@ -4,7 +4,7 @@ import core.ReceivePacket;
 
 import java.io.ByteArrayOutputStream;
 public abstract class AbsByteArrayReceivePacket<Entity> extends ReceivePacket<ByteArrayOutputStream,Entity> {
-    public AbsByteArrayReceivePacket(int len) {
+    public AbsByteArrayReceivePacket(long len) {
         super(len);
     }
